# [Refactor] Faster creation on device