--- a/tensordict/tensordict.py
+++ b/tensordict/tensordict.py
@@ -1557,9 +1557,7 @@ class TensorDictBase(MutableMapping):
                     f"={_shape(value)[: self.batch_dims]} with value {value}"
                 )
 
-        if self.device is not None and self.device.type == "cuda" and value.device == torch.device("cpu"):
-            value = value.pin_memory().to(self.device, non_blocking=True)
-        else:
+        if self.device is not None:
             value = value.to(self.device, non_blocking=True)
 
         if (