We call `to(..., non_blocking=True)` in TD creation to make things faster

Before
```python
%%timeit
td = TensorDict({s: torch.zeros(100, 100) for s in string.ascii_lowercase}, [], device="cuda:0")
```
1.16 ms ± 279 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)

After
783 µs ± 167 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)


Conversely, before:
```python
%%timeit
td = TensorDict({s: torch.zeros(100, 100, device="cuda:0") for s in string.ascii_lowercase}, [], device="cpu")
```
1.25 ms ± 170 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)

after
613 µs ± 27 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)

For the record:
I tested pin_memory but it does not help