@vmoens 2023-05-19
This is due to the problem I mentioned yesterday that no benchmark is actually running on gpu :( 
I tried a couple of things [there](#388) without success.
If you want to have a look ;)