@vmoens 2023-05-22
In general I think it's safe
The only thing that this does it to let us load all the data asynchronously on device, but to access it you'll have to wait for it to be there...